+++ deploy/sdk/src/dynamo/sdk/__init__.py
@@ -17,8 +17,8 @@ from typing import Any
 
 from bentoml import on_shutdown as async_on_shutdown
 
-from dynamo.sdk.core.decorators.endpoint import dynamo_api, dynamo_endpoint
-from dynamo.sdk.core.lib import DYNAMO_IMAGE, depends, service
+from dynamo.core_sdk.decorators.endpoint import dynamo_api, dynamo_endpoint
+from dynamo.core_sdk.lib import DYNAMO_IMAGE, depends, service
 from dynamo.sdk.lib.decorators import async_on_start
 
 dynamo_context: dict[str, Any] = {}
