+++ lib/bindings/python/src/dynamo/core_sdk/runner/dynamo.py
@@ -28,8 +28,8 @@ from circus.sockets import CircusSocket
 from circus.watcher import Watcher
 from fastapi import FastAPI
 
-from dynamo.sdk.core.decorators.endpoint import DynamoClient, DynamoEndpoint
-from dynamo.sdk.core.protocol.interface import (
+from dynamo.core_sdk.decorators.endpoint import DynamoClient, DynamoEndpoint
+from dynamo.core_sdk.protocol.interface import (
     DependencyInterface,
     DeploymentTarget,
     DynamoConfig,
@@ -39,7 +39,7 @@ from dynamo.sdk.core.protocol.interface import (
     ServiceConfig,
     ServiceInterface,
 )
-from dynamo.sdk.core.runner.common import ServiceMixin
+from dynamo.core_sdk.runner.common import ServiceMixin
 
 logger = logging.getLogger(__name__)
